Move SearchFilters into internal/search/streaming
